job_search_ai/requirements.txt
crewai==0.1.25
langchain==0.1.12
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.1
numpy==1.26.4
aiohttp==3.9.3